requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
psycopg2-binary  
python-dotenv    
gunicorn         
Flask-Migrate==4.1.0
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
python-dotenv==1.1.0
requests==2.32.3
Werkzeug==3.1.3
alembic==1.15.2
Flask-JWT-Extended==4.7.1
flask-cors==5.0.1